scripts/old_versions/requirements_docker.txt
# Docker environment requirements for training
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
transformers==4.35.2
diffusers==0.23.1
accelerate==0.24.1
huggingface-hub==0.19.4
safetensors==0.4.1
peft==0.6.2
bitsandbytes==0.41.3
xformers==0.0.22
tensorboard==2.15.1
wandb==0.16.1
pillow==10.1.0
scipy==1.11.4
numpy==1.24.4
tqdm==4.66.1
matplotlib==3.8.2
soundfile==0.12.1
librosa==0.10.1
omegaconf==2.3.0
gradio==4.8.0